+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 =================
 
 ## 0.18.46
+### is.codion.common.db
+- AbstractConnectionPoolWrapper now logs check out times in microseconds.
 ### is.codion.swing.common.model
 - FilterTableModel.TableColumns.caption() added.
 - FilterTableModel.TableColumns.description() added.